- AutoRemoteSyscalls::{syscall, syscall_base} - Remove Deref/DerefMut for dyn TaskInterface - Add open_mem_fd() to the TaskInterface - Move open_mem_fd() out of impl Task as an independent method:     open_mem_fd<T:TaskInterface>() - Call open_mem_fd<T> from RecordTask::open_mem_fd() - Misc changes; fix various borrow issues

+++ src/registers.rs
@@ -764,7 +764,7 @@ impl Registers {
         }
     }
 
-    pub fn set_arg(&mut self, index: i32, value: usize) {
+    pub fn set_arg(&mut self, index: usize, value: usize) {
         match index {
             1 => self.set_arg1(value),
             2 => self.set_arg2(value),
